// constants/places.enum.ts
export enum PlaceType {
    HOSPITAL = 'hospital',
    SCHOOL = 'school',
    RESTAURANT = 'restaurant',
    BUS_STOP = 'bus_stop',
    HOTEL = 'hotel',
    POLICE_STATION = 'police_station',
    PHARMACY = 'pharmacy',
    BANK = 'bank',
    GAS_STATION = 'gas_station',
    SUPERMARKET = 'supermarket',
    PARK = 'park',
    CHURCH = 'church',
    POST_OFFICE = 'post_office',
    LIBRARY = 'library',
    MUSEUM = 'museum',
    CINEMA = 'cinema',
    SHOPPING_MALL = 'shopping_mall',
    AIRPORT = 'airport',
    TRAIN_STATION = 'train_station',
    SUBWAY_STATION = 'subway_station',
    TAXI_STAND = 'taxi_stand',
    PARKING = 'parking',
    ATM = 'atm',
    FIRE_STATION = 'fire_station',
    VETERINARY = 'veterinary',
    DENTIST = 'dentist',
    CLINIC = 'clinic',
    CAFE = 'cafe',
    BAR = 'bar',
    UNIVERSITY = 'university',
    KINDERGARTEN = 'kindergarten',
  }
  